image_filters/requirements.txt
numpy
pillow
opencv-python-headless
streamlit